+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 2017-03-01 - v0.9.5
 
  - inline typeahead works now with an array and a function
+ - inline typeahead accepts an display property value
 
 ### 2017-01-31 - v0.9.4